submission history component structure
